pyproject.toml
[tool.poetry]
name = "roomsense2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
boto3 = "^1.28.62"
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
motor = "^3.3.1"
tomli = "^2.0.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
